+++ packages/2019-disaster-game/src/components/Game/index.js
@@ -31,8 +31,7 @@ import DefaultScreen from "./DefaultScreen/index";
 import AttractorScreen from "./AttractorScreen/index";
 import KitScreen from "./KitScreen/index";
 import QuakeScreen from "./QuakeScreen/index";
-// import TaskScreen from "./TaskScreen/index";
-import TaskScreen from "./TaskScreen/index.new";
+import TaskScreen from "./TaskScreen/index";
 import SummaryScreen from "./SummaryScreen/index";
 import BetweenScreen from "./BetweenScreen/index";
 import KitIntro from "./BetweenScreen/KitIntro";
